+++ autosklearn/automl.py
@@ -8,9 +8,7 @@ import multiprocessing
 import os
 import sys
 import time
-from typing import (
-    Any, Dict, Optional, List, Tuple, Union, Iterable, Final, Literal
-)
+from typing import Any, Dict, Optional, List, Tuple, Union
 import uuid
 import unittest.mock
 import warnings
@@ -1560,213 +1558,6 @@ class AutoML(BaseEstimator):
         sio.write('  Number of target algorithms that exceeded the memory '
                   'limit: %d\n' % num_memout)
         return sio.getvalue()
-
-    def leaderboard(
-        self,
-        detailed: bool = False,
-        ensemble_only: bool = True,
-        top_k: Union[int, Literal['all']] = 'all',
-        sort_by: Literal['cost', 'ensemble_weight'] = 'cost',
-        include: Optional[Iterable[str]] = None
-    ) -> pd.DataFrame:
-        """ Returns a pandas table of results for all evaluated models.
-
-        Gives an overview of all models trained during the search process along
-        with various statistics about their training.
-
-        The availble statistics are:
-
-            Simple:
-                * id - The id used to identify the model.
-                * rank - The rank of the model.
-                * ensemble_weight - The weight given to the model in the
-                ensemble.
-                * type - The type of classifier/regressor used.
-                * cost - The loss of the model on the validation set.
-                * duration - Length of time the model was optimized for.
-
-            Includable:
-                * budget - How much time was allocated to this model.
-                * status - The return status of training the model with SMAC.
-                * train_loss - The loss of the model on the training set.
-                * balancing_strategy - TODO
-                * start_time - Time the model began being optimized
-                * end_time - Time the model ended being optimized
-                * data_preprocessors - The preprocessors used on the data
-                * feature_preprocessors - The preprocessors for features types
-
-        Paramaters
-        ----------
-        detailed: bool = False
-            Whether to give detailed information or just a simple overview.
-
-        ensemble_only: bool = True
-            Whether to view only models included in the ensemble or all models
-            trained.
-
-        top_k: int or 'all'
-            How many models to display.
-
-        sort_by: 'cost' or 'ensemble_weight'
-            Whether to sort model ranks by 'cost' or
-
-        include: Optional[Iterable[str]] = None
-            Items to include, other items not specified will be excluded.
-            If left as `None`, it will resort back to the the `simple` or
-            `detailed` view as specified by the `detailed` parameter.
-
-        Returns
-        -------
-        pd.DataFrame
-            A dataframe of statistics for the models, ordered by their rank.
-        """
-        # Validation
-        valid_sort_by = ['cost', 'ensemble_weight']
-        if sort_by not in valid_sort_by:
-            raise ValueError(f"sort_by='{sort_by}' must be one of "
-                             f"{valid_sort_by}")
-
-        if (
-            not (isinstance(top_k, str) or isinstance(top_k, int))
-            or (isinstance(top_k, str) and top_k != 'all')
-            or (isinstance(top_k, int) and top_k <= 0)
-        ):
-            raise ValueError(f"top_k={top_k} must be a positive integer or pass"
-                             f"'all' to view all results")
-
-        # TODO budget seems to be 0.0 all the time?
-        # TODO validate that `self` is fitted. This is required for
-        #      self.ensemble_ to get the identifiers of models it will generate
-        #      weights for.
-
-        # The different kinds of column ouputs based on the parameters
-        all_columns: Final[List[str]] = [
-            "id", "rank", "ensemble_weight", "type", "cost", "duration",
-            "train_loss", "seed", "start_time", "end_time", "budget", "status",
-            "data_preprocessors", "feature_preprocessors", "balancing_strategy",
-            "config_origin"
-        ]
-        simple_columns: Final[List[str]] = [
-            "id", "rank", "ensemble_weight", "type", "cost", "duration"
-        ]
-        detailed_columns: Final[List[str]] = all_columns
-
-        # Decide which columns to include in the output
-        if include is not None:
-            columns = list(include)  # Incase `include` is generator
-
-            invalid_include_items = set(columns) - set(all_columns)
-            if len(invalid_include_items) != 0:
-                raise ValueError(f"Values {invalid_include_items} are not known"
-                                 f" columns to include, must be contained in "
-                                 f"{all_columns}")
-        elif detailed:
-            columns = detailed_columns
-        else:
-            columns = simple_columns
-
-        # To get all the models that were optmized, we collect what we can from
-        # runhistory first.
-        def has_key(rv, key):
-            return rv.additional_info and key in rv.additional_info
-
-        model_runs = {
-            rkey.config_id: {
-                'id': rkey.config_id,
-                'seed': rkey.seed,
-                'budget': rkey.budget,
-                'duration': rval.time,
-                'start_time': rval.starttime,
-                'end_time': rval.endtime,
-                'status': str(rval.status),
-                'cost': rval.cost,
-                'train_loss': rval.additional_info['train_loss']
-                if has_key(rval, 'train_loss') else None,
-                'config_origin': rval.additional_info['configuration_origin']
-                if has_key(rval, 'configuration_origin') else None
-            }
-            for rkey, rval in self.runhistory_.data.items()
-        }
-
-        # Next we get some info about the model itself
-        # TODO 78xh - Using the `is_classification` passed to `fit()` for now.
-        #      Should really be an abstract property of the subclass?
-        type_attr_str = 'classifier' if self._is_classificaiton else 'regressor'
-
-        # A dict mapping model ids to their configurations
-        configurations = self.runhistory_.ids_config
-
-        for model_id, run_info in model_runs.items():
-            run_config = configurations[model_id]._values
-
-            run_info.update({
-                'balancing_strategy': run_config['balancing:strategy'],
-                'type': run_config[f'{type_attr_str}:__choice__'],
-                'data_preprocessors': [
-                    value for key, value in run_config.items()
-                    if 'data_preprocessing' in key and '__choice__' in key
-                ],
-                'feature_preprocessors': [
-                    value for key, value in run_config.items()
-                    if 'feature_preprocessor' in key and '__choice__' in key
-                ]
-            })
-
-        # Get the models ensemble weight if it has one
-        # TODO both implementing classes of AbstractEnsemble have a property
-        #      `identifiers_` and `weights_`,
-        #       might be good to put it as an abstract property
-        # TODO `ensemble_.identifiers_` and `ensemble_.weights_` are loosely
-        #      tied together by ordering, might be better to store as tuple
-        for i, (_, model_id, _) in enumerate(self.ensemble_.identifiers_):
-            model_runs[model_id]['ensemble_weight'] = self.ensemble_.weights_[i]
-
-        # Filter out non-ensemble members if needed, else fill in a default
-        # value of 0 if it's missing
-        if ensemble_only:
-            model_runs = {
-                model_id: info
-                for model_id, info in model_runs.items()
-                if ('ensemble_weight' in info and info['ensemble_weight'] > 0)
-            }
-        else:
-            for model_id, info in model_runs.items():
-                if 'ensemble_weight' not in info:
-                    info['ensemble_weight'] = 0
-
-        # Finally, convert into a tabular format by converting the dict into
-        # column wise orientation.
-        # We don't have `rank` yet so we ignore it for now
-        dataframe = pd.DataFrame({
-            col: [run_info[col] for run_info in model_runs.values()]
-            for col in columns if col != 'rank'
-        })
-
-        # Sort the values of the specified column
-        column_sort_ascending = {
-            'ensemble_weight': False,
-            'cost': True
-        }
-        dataframe.sort_values(by=sort_by,
-                              ascending=column_sort_ascending[sort_by],
-                              inplace=True)
-
-        # Give it an index
-        dataframe.set_index('id', inplace=True)
-
-        # Add the rank column at the position given by columns
-        if 'rank' in columns:
-            dataframe.insert(column='rank',
-                             value=range(1, len(dataframe)),
-                             loc=list(columns).index('rank'))
-
-        # Lastly, just grab the top_k
-        if top_k == 'all' or top_k >= len(dataframe):
-            top_k = len(dataframe) - 1
-
-        dataframe = dataframe.head(top_k)
-
-        return dataframe
 
     def get_models_with_weights(self):
         if self.models_ is None or len(self.models_) == 0 or \
